gravel: nodes: spin-off etcd functions to module
